requirements.txt
haystack-ai
kafka-python
azure-storage-blob
python-dotenv
pydantic
